requirements.txt
flask
flask-cors
requests
apscheduler
fpdf
pywhatkit
bcrypt
PyJWT
pywebview
